backend/fastapi_vision/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.1
numpy>=1.24.3
easyocr>=1.7.0
opencv-python>=4.8.1.78 